# FileStream optimizations