--- a/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/LegacyFileStreamStrategy.Windows.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/LegacyFileStreamStrategy.Windows.cs
@@ -1037,15 +1037,7 @@ namespace System.IO.Strategies
                     }
 
                     completionSource.ReleaseNativeResource();
-
-                    if (errorCode == ERROR_HANDLE_EOF)
-                    {
-                        ThrowHelper.ThrowEndOfFileException();
-                    }
-                    else
-                    {
-                        throw Win32Marshal.GetExceptionForWin32Error(errorCode, _path);
-                    }
+                    throw Win32Marshal.GetExceptionForWin32Error(errorCode, _path);
                 }
                 else if (cancellationToken.CanBeCanceled) // ERROR_IO_PENDING
                 {