--- a/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/LegacyFileStreamStrategy.Windows.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/LegacyFileStreamStrategy.Windows.cs
@@ -1037,7 +1037,15 @@ namespace System.IO.Strategies
                     }
 
                     completionSource.ReleaseNativeResource();
-                    throw Win32Marshal.GetExceptionForWin32Error(errorCode, _path);
+
+                    if (errorCode == ERROR_HANDLE_EOF)
+                    {
+                        ThrowHelper.ThrowEndOfFileException();
+                    }
+                    else
+                    {
+                        throw Win32Marshal.GetExceptionForWin32Error(errorCode, _path);
+                    }
                 }
                 else if (cancellationToken.CanBeCanceled) // ERROR_IO_PENDING
                 {