--- a/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/SyncWindowsFileStreamStrategy.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/Strategies/SyncWindowsFileStreamStrategy.cs
@@ -159,7 +159,7 @@ namespace System.IO.Strategies
             }
             Debug.Assert(r >= 0, "FileStream's WriteCore is likely broken.");
             _filePosition += r;
-            return;
+            UpdateLengthOnChangePosition();
         }
 
         private NativeOverlapped GetNativeOverlappedForCurrentPosition()