… src/libraries/System.IO.FileSystem/tests/FileStream/SafeFileHandle.cs
… src/libraries/System.Private.CoreLib/src/System/IO/Strategies/WindowsFileStreamStrategy.cs