This PR just merges optimizations made in https://github.com/dotnet/runtime/pull/49145 and https://github.com/dotnet/runtime/pull/49638 and rebases them in top of the latest changes recently introduced by @adamsitnik in https://github.com/dotnet/runtime/pull/48813.

Fixes https://github.com/dotnet/runtime/issues/49541.

I had to do one adjustment (https://github.com/dotnet/runtime/commit/64e5174524b8a6d96c0b0e101977638ac13fffc5) in order to satisfy the tests added by https://github.com/dotnet/runtime/pull/49754 but otherwise, everything else remains as it was in the old PRs.

Perf results: 
There is an allocation increase of 16 B I suspect is caused by the newly added fields `_share` and `_length`.

<details>

``` ini

BenchmarkDotNet=v0.12.1.1521-nightly, OS=Windows 10.0.19042.868 (20H2/October2020Update)
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=6.0.100-preview.2.21153.28
  [Host]     : .NET 6.0.0 (6.0.21.12307), X64 RyuJIT
  Job-IRUTZB : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT
  Job-LEWZXC : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                 Method | Toolchain |  fileSize | userBufferSize |      options |            Mean |          Error |         StdDev |          Median |             Min |             Max | Ratio |   Allocated |
|----------------------- |---------- |---------- |--------------- |------------- |----------------:|---------------:|---------------:|----------------:|----------------:|----------------:|------:|------------:|
|                   Read | base      |      1024 |           1024 |         None |        42.15 μs |       0.409 μs |       0.342 μs |        42.05 μs |        41.67 μs |        42.89 μs |  1.00 |     4,328 B |
|                   Read | feature   |      1024 |           1024 |         None |        43.11 μs |       0.918 μs |       1.057 μs |        43.01 μs |        41.61 μs |        45.08 μs |  1.01 |     4,344 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|                  Write | base      |      1024 |           1024 |         None |     1,034.05 μs |      21.974 μs |      22.565 μs |     1,038.34 μs |       976.98 μs |     1,073.31 μs |  1.00 |     4,328 B |
|                  Write | feature   |      1024 |           1024 |         None |     1,042.44 μs |      39.041 μs |      44.960 μs |     1,045.14 μs |       965.48 μs |     1,130.16 μs |  1.01 |     4,344 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              ReadAsync | base      |      1024 |           1024 |         None |        86.86 μs |       2.158 μs |       2.485 μs |        86.53 μs |        83.21 μs |        92.46 μs |  1.00 |     5,072 B |
|              ReadAsync | feature   |      1024 |           1024 |         None |        86.38 μs |       1.321 μs |       1.236 μs |        85.92 μs |        84.78 μs |        88.24 μs |  0.99 |     5,088 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             WriteAsync | base      |      1024 |           1024 |         None |     1,062.25 μs |      44.082 μs |      50.765 μs |     1,068.88 μs |       987.20 μs |     1,143.93 μs |  1.00 |     4,416 B |
|             WriteAsync | feature   |      1024 |           1024 |         None |     1,038.11 μs |      21.939 μs |      25.265 μs |     1,042.38 μs |       996.82 μs |     1,093.33 μs |  0.98 |     4,432 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              ReadAsync | base      |      1024 |           1024 | Asynchronous |       108.90 μs |       2.082 μs |       2.045 μs |       108.18 μs |       106.64 μs |       113.34 μs |  1.00 |     5,240 B |
|              ReadAsync | feature   |      1024 |           1024 | Asynchronous |       102.23 μs |       1.785 μs |       1.670 μs |       101.36 μs |       100.33 μs |       105.17 μs |  0.94 |     5,256 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             WriteAsync | base      |      1024 |           1024 | Asynchronous |     1,173.33 μs |      44.106 μs |      49.024 μs |     1,158.02 μs |     1,111.86 μs |     1,297.30 μs |  1.00 |     4,826 B |
|             WriteAsync | feature   |      1024 |           1024 | Asynchronous |     1,120.89 μs |      32.816 μs |      36.475 μs |     1,114.59 μs |     1,056.56 μs |     1,177.69 μs |  0.96 |     4,838 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              OpenClose | base      |      1024 |              ? |         None |        35.24 μs |       0.641 μs |       0.569 μs |        35.30 μs |        33.88 μs |        36.25 μs |  1.00 |       208 B |
|              OpenClose | feature   |      1024 |              ? |         None |        33.30 μs |       0.582 μs |       0.516 μs |        33.18 μs |        32.78 μs |        34.36 μs |  0.94 |       224 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             LockUnlock | base      |      1024 |              ? |         None |        47.73 μs |       1.236 μs |       1.374 μs |        48.01 μs |        45.64 μs |        50.13 μs |  1.00 |       208 B |
|             LockUnlock | feature   |      1024 |              ? |         None |        46.02 μs |       0.694 μs |       0.649 μs |        46.05 μs |        44.95 μs |        47.27 μs |  0.96 |       224 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|            SeekForward | base      |      1024 |              ? |         None |       679.71 μs |      12.643 μs |      11.826 μs |       677.27 μs |       663.22 μs |       699.84 μs |  1.00 |       208 B |
|            SeekForward | feature   |      1024 |              ? |         None |        43.52 μs |       0.798 μs |       0.746 μs |        43.33 μs |        42.57 μs |        44.90 μs |  0.06 |       224 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|           SeekBackward | base      |      1024 |              ? |         None |     2,409.29 μs |      34.724 μs |      30.782 μs |     2,413.36 μs |     2,364.14 μs |     2,463.39 μs |  1.00 |       208 B |
|           SeekBackward | feature   |      1024 |              ? |         None |     1,837.29 μs |      33.029 μs |      30.895 μs |     1,826.06 μs |     1,801.63 μs |     1,891.95 μs |  0.76 |       224 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|               ReadByte | base      |      1024 |              ? |         None |        43.99 μs |       0.570 μs |       0.505 μs |        43.83 μs |        43.33 μs |        45.06 μs |  1.00 |     4,328 B |
|               ReadByte | feature   |      1024 |              ? |         None |        44.32 μs |       0.822 μs |       0.729 μs |        44.12 μs |        43.43 μs |        45.92 μs |  1.01 |     4,344 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              WriteByte | base      |      1024 |              ? |         None |     1,107.18 μs |      27.937 μs |      29.892 μs |     1,116.16 μs |     1,051.06 μs |     1,141.02 μs |  1.00 |     4,328 B |
|              WriteByte | feature   |      1024 |              ? |         None |     1,147.78 μs |      50.184 μs |      57.792 μs |     1,134.50 μs |     1,058.34 μs |     1,282.50 μs |  1.04 |     4,344 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|                  Flush | base      |      1024 |              ? |         None |     4,671.56 μs |     268.251 μs |     308.919 μs |     4,777.24 μs |     4,159.09 μs |     5,111.56 μs |  1.00 |     4,329 B |
|                  Flush | feature   |      1024 |              ? |         None |     4,739.84 μs |     299.293 μs |     344.666 μs |     4,916.42 μs |     4,233.60 μs |     5,190.47 μs |  1.02 |     4,345 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             FlushAsync | base      |      1024 |              ? |         None |     6,215.18 μs |     323.164 μs |     372.156 μs |     6,392.51 μs |     5,676.52 μs |     6,679.63 μs |  1.00 |   275,049 B |
|             FlushAsync | feature   |      1024 |              ? |         None |     6,268.75 μs |     355.372 μs |     409.247 μs |     6,446.29 μs |     5,754.60 μs |     6,804.16 μs |  1.01 |   275,065 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             CopyToFile | base      |      1024 |              ? |         None |     1,106.93 μs |      39.110 μs |      43.471 μs |     1,107.62 μs |     1,047.41 μs |     1,196.00 μs |  1.00 |     4,538 B |
|             CopyToFile | feature   |      1024 |              ? |         None |     1,150.55 μs |      34.885 μs |      38.775 μs |     1,156.79 μs |     1,088.84 μs |     1,213.33 μs |  1.04 |     4,569 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|        CopyToFileAsync | base      |      1024 |              ? |         None |     1,267.53 μs |      45.676 μs |      52.601 μs |     1,266.82 μs |     1,192.55 μs |     1,349.69 μs |  1.00 |     5,562 B |
|        CopyToFileAsync | feature   |      1024 |              ? |         None |     1,262.80 μs |      39.062 μs |      41.796 μs |     1,269.05 μs |     1,188.39 μs |     1,347.70 μs |  1.00 |     5,596 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              OpenClose | base      |      1024 |              ? | Asynchronous |        34.54 μs |       0.473 μs |       0.419 μs |        34.43 μs |        33.91 μs |        35.57 μs |  1.00 |       256 B |
|              OpenClose | feature   |      1024 |              ? | Asynchronous |        34.97 μs |       0.652 μs |       0.610 μs |        34.79 μs |        34.12 μs |        35.99 μs |  1.01 |       272 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             LockUnlock | base      |      1024 |              ? | Asynchronous |        49.49 μs |       0.932 μs |       0.915 μs |        49.15 μs |        48.32 μs |        51.42 μs |  1.00 |       256 B |
|             LockUnlock | feature   |      1024 |              ? | Asynchronous |        50.07 μs |       0.985 μs |       0.873 μs |        49.91 μs |        49.18 μs |        51.92 μs |  1.01 |       272 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|            SeekForward | base      |      1024 |              ? | Asynchronous |     2,621.18 μs |      33.964 μs |      31.770 μs |     2,631.60 μs |     2,576.36 μs |     2,670.52 μs |  1.00 |       256 B |
|            SeekForward | feature   |      1024 |              ? | Asynchronous |        44.94 μs |       0.763 μs |       0.714 μs |        44.62 μs |        44.00 μs |        46.49 μs |  0.02 |       272 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|           SeekBackward | base      |      1024 |              ? | Asynchronous |     5,290.45 μs |      91.791 μs |      85.861 μs |     5,277.68 μs |     5,169.49 μs |     5,421.50 μs |  1.00 |       257 B |
|           SeekBackward | feature   |      1024 |              ? | Asynchronous |     2,678.27 μs |      46.203 μs |      43.218 μs |     2,659.78 μs |     2,621.72 μs |     2,748.75 μs |  0.51 |       272 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|               ReadByte | base      |      1024 |              ? | Asynchronous |        67.08 μs |       1.265 μs |       1.406 μs |        66.66 μs |        65.48 μs |        69.48 μs |  1.00 |     4,698 B |
|               ReadByte | feature   |      1024 |              ? | Asynchronous |        61.55 μs |       1.344 μs |       1.547 μs |        61.35 μs |        59.63 μs |        64.36 μs |  0.92 |     4,715 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              WriteByte | base      |      1024 |              ? | Asynchronous |     1,232.27 μs |      34.232 μs |      39.421 μs |     1,229.81 μs |     1,177.33 μs |     1,311.58 μs |  1.00 |     4,727 B |
|              WriteByte | feature   |      1024 |              ? | Asynchronous |     1,199.37 μs |      43.552 μs |      50.155 μs |     1,195.56 μs |     1,127.21 μs |     1,292.33 μs |  0.97 |     4,755 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|                  Flush | base      |      1024 |              ? | Asynchronous |    47,875.35 μs |   2,339.326 μs |   2,693.971 μs |    47,131.61 μs |    44,045.60 μs |    53,199.18 μs |  1.00 |   155,908 B |
|                  Flush | feature   |      1024 |              ? | Asynchronous |    14,945.85 μs |     494.672 μs |     569.665 μs |    14,670.86 μs |    14,347.10 μs |    16,130.53 μs |  0.31 |   152,444 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             FlushAsync | base      |      1024 |              ? | Asynchronous |    65,640.53 μs |   3,328.732 μs |   3,699.877 μs |    66,847.73 μs |    57,764.97 μs |    69,690.27 μs |  1.00 |   307,926 B |
|             FlushAsync | feature   |      1024 |              ? | Asynchronous |    17,211.85 μs |     575.562 μs |     662.818 μs |    17,156.76 μs |    16,383.39 μs |    18,551.23 μs |  0.26 |   307,947 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|        CopyToFileAsync | base      |      1024 |              ? | Asynchronous |     1,444.81 μs |      48.168 μs |      55.470 μs |     1,452.91 μs |     1,310.17 μs |     1,533.31 μs |  1.00 |     6,219 B |
|        CopyToFileAsync | feature   |      1024 |              ? | Asynchronous |     1,838.93 μs |     251.001 μs |     278.987 μs |     1,855.54 μs |     1,358.64 μs |     2,233.14 μs |  1.27 |     6,191 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|                   Read | base      |   1048576 |            512 |         None |     1,054.38 μs |     137.667 μs |     158.537 μs |       984.19 μs |       896.97 μs |     1,362.93 μs |  1.00 |     4,328 B |
|                   Read | feature   |   1048576 |            512 |         None |       821.17 μs |      16.453 μs |      18.287 μs |       816.62 μs |       796.64 μs |       864.23 μs |  0.79 |     4,344 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|                  Write | base      |   1048576 |            512 |         None |     6,409.50 μs |     261.978 μs |     280.314 μs |     6,381.70 μs |     6,035.65 μs |     7,055.64 μs |  1.00 |     4,329 B |
|                  Write | feature   |   1048576 |            512 |         None |     6,052.51 μs |     206.714 μs |     238.052 μs |     6,025.02 μs |     5,631.29 μs |     6,402.86 μs |  0.94 |     4,345 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|              ReadAsync | base      |   1048576 |            512 |         None |     1,660.33 μs |     103.631 μs |     119.342 μs |     1,696.76 μs |     1,479.36 μs |     1,877.67 μs |  1.00 |    86,671 B |
|              ReadAsync | feature   |   1048576 |            512 |         None |     1,493.91 μs |      33.582 μs |      38.673 μs |     1,480.58 μs |     1,439.02 μs |     1,580.41 μs |  0.90 |    86,686 B |
|                        |           |           |                |              |                 |                |                |                 |                 |                 |       |             |
|             WriteAsync | base      |   1048576 |            512 |         None |     8,570.25 μs |     503.775 μs |     580.147 μs |     8,503.81 μs |     7,755.54 μs |     9,810.12 μs |  1.00 |    78,194 B |
|             WriteAsync | feature   |   1048576 |            512 |         None |     7,507.00 μs |     271.475 μs |     312.631 μs |     7,496.33 μs |     7,008.36 μs |     8,183.34 μs |  0.88 |    78,313 B |
…