@Jozkee 2021-03-24
> Could you please re-run the following benchmarks:

@adamsitnik I ran the suggested benchmarks again and the results show no regression nor improvement vs base (main), it seems to me that the benchmarks are a bit unstable. 

Read/Write sync:
```
| Method |           Toolchain | fileSize | userBufferSize | options |       Mean |     Error |    StdDev |     Median |        Min |        Max | Ratio | Allocated |
|------- |-------------------- |--------- |--------------- |-------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|------:|----------:|
|   Read |   \base\CoreRun.exe |  1048576 |            512 |    None |   841.1 us |   7.18 us |   6.00 us |   840.9 us |   833.2 us |   854.1 us |  1.00 |      4 KB |
|   Read |\feature\CoreRun.exe |  1048576 |            512 |    None |   842.1 us |   8.90 us |   8.32 us |   843.6 us |   826.5 us |   857.3 us |  1.00 |      4 KB |
|        |                     |          |                |         |            |           |           |            |            |            |       |           |
|  Write |   \base\CoreRun.exe |  1048576 |            512 |    None | 6,825.0 us | 350.07 us | 403.15 us | 6,736.7 us | 6,310.4 us | 7,601.0 us |  1.00 |      4 KB |
|  Write |\feature\CoreRun.exe |  1048576 |            512 |    None | 6,834.6 us | 203.73 us | 226.44 us | 6,827.1 us | 6,507.2 us | 7,296.0 us |  1.01 |      4 KB |
```

CopyToFileAsync 1024 file size:
```
|          Method |           Toolchain |  fileSize |      options |       Mean |      Error |     StdDev |     Median |        Min |        Max | Ratio | Allocated |
|---------------- |-------------------- |---------- |------------- |-----------:|-----------:|-----------:|-----------:|-----------:|-----------:|------:|----------:|
| CopyToFileAsync |   \base\CoreRun.exe |      1024 |         None |   1.561 ms |  0.0748 ms |  0.0861 ms |   1.543 ms |   1.427 ms |   1.742 ms |  1.00 |      5 KB |
| CopyToFileAsync |\feature\CoreRun.exe |      1024 |         None |   1.586 ms |  0.0464 ms |  0.0534 ms |   1.576 ms |   1.486 ms |   1.696 ms |  1.02 |      5 KB |
|                 |                     |           |              |            |            |            |            |            |            |       |           |
| CopyToFileAsync |   \base\CoreRun.exe |      1024 | Asynchronous |   1.847 ms |  0.1436 ms |  0.1596 ms |   1.853 ms |   1.597 ms |   2.099 ms |  1.00 |      6 KB |
| CopyToFileAsync |\feature\CoreRun.exe |      1024 | Asynchronous |   1.747 ms |  0.0747 ms |  0.0831 ms |   1.768 ms |   1.626 ms |   1.951 ms |  0.95 |      6 KB |
```